refactor: 🛠  moved icons to a common folder
